+++ internal/presentation/rest/handlers/withdraw/schema.go
@@ -13,8 +13,8 @@ func (h) GetMethod() string {
 }
 
 type request struct {
-	Order string `json:"order" validate:"required"`
-	Sum   int    `json:"sum" validate:"required,gt=0"`
+	Order string  `json:"order" validate:"required"`
+	Sum   float64 `json:"sum" validate:"required,gt=0"`
 }
 
 func (h) Parse(ctx *gin.Context) (request, error) {
