+++ internal/repository/withdrawals/interfaces.go
@@ -8,6 +8,6 @@ import (
 
 type WithdrawalsStorage interface {
 	FindWithdrawalsByUserID(ctx context.Context, userID int) ([]models.Withdrawal, error)
-	InsertWithdrawal(ctx context.Context, userID int, order string, sum int) error
-	SumWithdrawn(ctx context.Context, userID int) (int, error)
+	InsertWithdrawal(ctx context.Context, userID int, order string, sum float64) error
+	SumWithdrawn(ctx context.Context, userID int) (float64, error)
 }
